Moved files into correct "address" namespace from "fraud"
