+++ lib/src/main/java/uk/gov/di/ipv/cri/address/library/annotations/ExcludeFromGeneratedCoverageReport.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.cri.fraud.library.annotations;
+package uk.gov.di.ipv.cri.address.library.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
