+++ lib/src/main/java/uk/gov/di/ipv/cri/address/library/exception/AccessTokenProcessingException.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.cri.fraud.library.exception;
+package uk.gov.di.ipv.cri.address.library.exception;
 
 public class AccessTokenProcessingException extends Exception {
     public AccessTokenProcessingException() {
