+++ lib/src/main/java/uk/gov/di/ipv/cri/address/library/exception/AccessTokenValidationException.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.cri.fraud.library.exception;
+package uk.gov.di.ipv.cri.address.library.exception;
 
 public class AccessTokenValidationException extends Exception {
     public AccessTokenValidationException() {
