+++ lib/src/main/java/uk/gov/di/ipv/cri/address/library/service/AccessTokenService.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.cri.fraud.library.service;
+package uk.gov.di.ipv.cri.address.library.service;
 
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
 import com.nimbusds.oauth2.sdk.AccessTokenResponse;
@@ -16,12 +16,12 @@ import com.nimbusds.oauth2.sdk.token.Tokens;
 import com.nimbusds.oauth2.sdk.util.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import uk.gov.di.ipv.cri.fraud.library.annotations.ExcludeFromGeneratedCoverageReport;
-import uk.gov.di.ipv.cri.fraud.library.exception.AccessTokenProcessingException;
-import uk.gov.di.ipv.cri.fraud.library.exception.AccessTokenValidationException;
-import uk.gov.di.ipv.cri.fraud.library.persistence.DataStore;
-import uk.gov.di.ipv.cri.fraud.library.persistence.item.AccessTokenItem;
-import uk.gov.di.ipv.cri.fraud.library.validation.ValidationResult;
+import uk.gov.di.ipv.cri.address.library.annotations.ExcludeFromGeneratedCoverageReport;
+import uk.gov.di.ipv.cri.address.library.exception.AccessTokenProcessingException;
+import uk.gov.di.ipv.cri.address.library.exception.AccessTokenValidationException;
+import uk.gov.di.ipv.cri.address.library.persistence.DataStore;
+import uk.gov.di.ipv.cri.address.library.persistence.item.AccessTokenItem;
+import uk.gov.di.ipv.cri.address.library.validation.ValidationResult;
 
 import java.net.URI;
 import java.time.Instant;
